+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.4
 - added support for copying notes from a file manager to QOwnNotes via drag and 
   drop (for [Issue #100](https://github.com/pbek/QOwnNotes/issues/100))
+- added more Hungarian translation (thank you Fabijenna)
 
 ## 1.1.3
 - added **Hungarian translation** (a big thank you to Fabijenna!)